Warn on deleting clusters with a user prompt.
